requirements.txt
SpeechRecognition==3.10.1
dotenv
pyttsx3
Pillow